// src/Game/types.ts
export interface GridInput {
  id: number;
  value: string;
}

export interface LinesToShow {
  column1: boolean;
  column2: boolean;
  row1: boolean;
  row2: boolean;
}

export interface OverlayToShow {
  overlay1: boolean;
  overlay2: boolean;
  overlay3: boolean;
  overlay4: boolean;
  overlay5: boolean;
  overlay6: boolean;
  overlay7: boolean;
  overlay8: boolean;
}

export type WinningIndex = [number, number, number];